// packages/domains/src/issues/enums.ts
export enum IssueType {
    TASK = 'task',
    BUG = 'bug',
    STORY = 'story',
    EPIC = 'epic',
    SUBTASK = 'subtask'
}

export enum IssueStatus {
    TODO = 'todo',
    IN_PROGRESS = 'in_progress',
    IN_REVIEW = 'in_review',
    DONE = 'done'
}

export enum IssuePriority {
    LOWEST = 'lowest',
    LOW = 'low',
    MEDIUM = 'medium',
    HIGH = 'high',
    HIGHEST = 'highest'
}